move redeem + claim functionality behind single modal
