+++ src/components/Dashboard/Rewards.tsx
@@ -1,31 +1,20 @@
 import { BigNumber } from '@ethersproject/bignumber'
-import { Button, Descriptions, Space, Statistic } from 'antd'
-import Modal from 'antd/lib/modal/Modal'
+import { Button, Descriptions, Modal, Space, Statistic } from 'antd'
 import ConfirmUnstakeTokensModal from 'components/modals/ConfirmUnstakeTokensModal'
 import ParticipantsModal from 'components/modals/ParticipantsModal'
-import CurrencySymbol from 'components/shared/CurrencySymbol'
+import RedeemModal from 'components/modals/RedeemModal'
 import FormattedAddress from 'components/shared/FormattedAddress'
-import InputAccessoryButton from 'components/shared/InputAccessoryButton'
-import FormattedNumberInput from 'components/shared/inputs/FormattedNumberInput'
 import { NetworkContext } from 'contexts/networkContext'
 import { ProjectContext } from 'contexts/projectContext'
 import { ThemeContext } from 'contexts/themeContext'
-import { UserContext } from 'contexts/userContext'
 import { constants } from 'ethers'
 import useContractReader, { ContractUpdateOn } from 'hooks/ContractReader'
 import { useErc20Contract } from 'hooks/Erc20Contract'
 import { OperatorPermission, useHasPermission } from 'hooks/HasPermission'
 import { ContractName } from 'models/contract-name'
-import { BallotState } from 'models/funding-cycle'
 import { useContext, useMemo, useState } from 'react'
 import { bigNumbersDiff } from 'utils/bigNumbersDiff'
-import {
-  formatPercent,
-  formattedNum,
-  formatWad,
-  fromWad,
-  parseWad,
-} from 'utils/formatNumber'
+import { formatPercent, formatWad } from 'utils/formatNumber'
 import { decodeFCMetadata } from 'utils/fundingCycle'
 
 import IssueTickets from './IssueTickets'
@@ -36,12 +25,13 @@ export default function Rewards({
 }: {
   totalOverflow: BigNumber | undefined
 }) {
-  const [stakeModalVisible, setStakeModalVisible] = useState<boolean>()
+  const [manageTokensModalVisible, setManageTokensModalVisible] = useState<
+    boolean
+  >()
   const [unstakeModalVisible, setUnstakeModalVisible] = useState<boolean>()
   const [participantsModalVisible, setParticipantsModalVisible] = useState<
     boolean
   >()
-  const { contracts, transactor } = useContext(UserContext)
   const { userAddress } = useContext(NetworkContext)
 
   const {
@@ -57,8 +47,6 @@ export default function Rewards({
   } = useContext(ThemeContext)
 
   const [redeemModalVisible, setRedeemModalVisible] = useState<boolean>(false)
-  const [redeemAmount, setRedeemAmount] = useState<string>()
-  const [loadingRedeem, setLoadingRedeem] = useState<boolean>()
 
   const ticketsUpdateOn: ContractUpdateOn = useMemo(
     () => [
@@ -107,25 +95,7 @@ export default function Rewards({
     updateOn: ticketsUpdateOn,
   })
 
-  const currentOverflow = useContractReader<BigNumber>({
-    contract: ContractName.TerminalV1,
-    functionName: 'currentOverflowOf',
-    args: projectId ? [projectId.toHexString()] : null,
-    valueDidChange: bigNumbersDiff,
-  })
-
-  const currentBallotState = useContractReader<BallotState>({
-    contract: ContractName.FundingCycles,
-    functionName: 'currentBallotStateOf',
-    args: projectId ? [projectId.toHexString()] : null,
-  })
-
   const metadata = decodeFCMetadata(currentFC?.metadata)
-
-  const bondingCurveRate =
-    currentBallotState === BallotState.Active
-      ? metadata?.reconfigurationBondingCurveRate
-      : metadata?.bondingCurveRate
 
   const reservedTicketBalance = useContractReader<BigNumber>({
     contract: ContractName.TerminalV1,
@@ -144,97 +114,9 @@ export default function Rewards({
     valueDidChange: bigNumbersDiff,
   })?.add(reservedTicketBalance ? reservedTicketBalance : BigNumber.from(0))
 
-  const base =
-    totalSupply && redeemAmount && currentOverflow
-      ? currentOverflow?.mul(parseWad(redeemAmount)).div(totalSupply)
-      : BigNumber.from(0)
-
-  const rewardAmount = useMemo(() => {
-    if (
-      !bondingCurveRate ||
-      !totalSupply ||
-      !base ||
-      !redeemAmount ||
-      !currentOverflow
-    )
-      return undefined
-
-    if (totalSupply.sub(parseWad(redeemAmount)).isNegative()) {
-      return currentOverflow
-    }
-
-    const number = base
-    const numerator = parseWad(bondingCurveRate).add(
-      parseWad(redeemAmount)
-        .mul(parseWad(200).sub(parseWad(bondingCurveRate)))
-        .div(totalSupply),
-    )
-    const denominator = parseWad(200)
-
-    return number.mul(numerator).div(denominator)
-  }, [redeemAmount, base, bondingCurveRate, totalSupply, currentOverflow])
-
   const totalBalance = iouBalance?.add(ticketsBalance ?? 0)
 
-  const maxClaimable = useContractReader<BigNumber>({
-    contract: ContractName.TerminalV1,
-    functionName: 'claimableOverflowOf',
-    args:
-      userAddress && projectId
-        ? [userAddress, projectId.toHexString(), totalBalance?.toHexString()]
-        : null,
-    valueDidChange: bigNumbersDiff,
-    updateOn: useMemo(
-      () =>
-        projectId && userAddress
-          ? [
-              {
-                contract: ContractName.TerminalV1,
-                eventName: 'Pay',
-                topics: [[], projectId.toHexString(), userAddress],
-              },
-              {
-                contract: ContractName.TerminalV1,
-                eventName: 'Redeem',
-                topics: [projectId.toHexString(), userAddress],
-              },
-            ]
-          : undefined,
-      [projectId],
-    ),
-  })
-
   const share = formatPercent(totalBalance, totalSupply)
-
-  function redeem() {
-    if (!transactor || !contracts || !rewardAmount) return
-
-    setLoadingRedeem(true)
-
-    const redeemWad = parseWad(redeemAmount)
-
-    if (!redeemWad || !projectId) return
-
-    // Arbitrary discrete value (wei) subtracted
-    const minAmount = rewardAmount?.sub(1e12).toHexString()
-
-    transactor(
-      contracts.TerminalV1,
-      'redeem',
-      [
-        userAddress,
-        projectId.toHexString(),
-        redeemWad.toHexString(),
-        minAmount,
-        userAddress,
-        false, // TODO preferconverted
-      ],
-      {
-        onConfirmed: () => setRedeemAmount(undefined),
-        onDone: () => setLoadingRedeem(false),
-      },
-    )
-  }
 
   const redeemDisabled = !totalOverflow || totalOverflow.eq(0)
 
@@ -306,7 +188,6 @@ export default function Rewards({
                       display: 'flex',
                       flexWrap: 'wrap',
                       justifyContent: 'space-between',
-                      alignItems: 'center',
                       width: '100%',
                     }}
                   >
@@ -326,27 +207,7 @@ export default function Rewards({
                       )}
                       <div>
                         {formatWad(iouBalance ?? 0, { decimals: 0 })}
-                        {ticketsIssued && (
-                          <>
-                            {' '}
-                            claimable{' '}
-                            <Button
-                              onClick={() => setUnstakeModalVisible(true)}
-                              type="text"
-                              size="small"
-                              disabled={iouBalance?.eq(0) || !iouBalance}
-                              style={
-                                iouBalance?.eq(0) || !iouBalance
-                                  ? {}
-                                  : {
-                                      color: colors.text.action.primary,
-                                    }
-                              }
-                            >
-                              Claim
-                            </Button>
-                          </>
-                        )}
+                        {ticketsIssued && <> claimable</>}
                       </div>
 
                       <div
@@ -360,13 +221,12 @@ export default function Rewards({
                         {share || 0}% of supply
                       </div>
                     </div>
+
                     <Button
-                      loading={loadingRedeem}
                       size="small"
-                      onClick={() => setRedeemModalVisible(true)}
-                      disabled={isPreviewMode}
+                      onClick={() => setManageTokensModalVisible(true)}
                     >
-                      Redeem
+                      Manage
                     </Button>
                   </div>
                 }
@@ -381,73 +241,36 @@ export default function Rewards({
       </Space>
 
       <Modal
-        title={`Redeem ${tokenSymbol ? tokenSymbol + ' tokens' : 'Tokens'}`}
+        title="Manage tokens"
+        visible={manageTokensModalVisible}
+        onCancel={() => setManageTokensModalVisible(false)}
+        okButtonProps={{ hidden: true }}
+      >
+        <Space direction="vertical" style={{ width: '100%' }}>
+          <Button
+            disabled={!ticketsIssued}
+            onClick={() => setUnstakeModalVisible(true)}
+            block
+          >
+            Claim tokens as ERC20
+          </Button>
+          <Button onClick={() => setRedeemModalVisible(true)} block>
+            Burn tokens for ETH
+          </Button>
+        </Space>
+      </Modal>
+      <RedeemModal
         visible={redeemModalVisible}
+        redeemDisabled={redeemDisabled}
+        totalBalance={totalBalance}
+        totalSupply={totalSupply}
         onOk={() => {
-          redeem()
           setRedeemModalVisible(false)
         }}
         onCancel={() => {
-          setRedeemAmount(undefined)
           setRedeemModalVisible(false)
         }}
-        okText={`Burn ${formattedNum(redeemAmount)} ${tokenSymbol ??
-          'tokens'} for ETH`}
-        okButtonProps={{
-          disabled:
-            redeemDisabled || !redeemAmount || parseInt(redeemAmount) === 0,
-        }}
-        width={540}
-      >
-        <Space direction="vertical" style={{ width: '100%' }}>
-          <div>
-            Balance: {formatWad(totalBalance ?? 0, { decimals: 0 })}{' '}
-            {tokenSymbol ?? 'tokens'}
-          </div>
-          <p>
-            Currently worth: <CurrencySymbol currency={0} />
-            {formatWad(maxClaimable, { decimals: 4 })}
-          </p>
-          <p>
-            Tokens can be redeemed for a portion of this project's ETH overflow,
-            according to the bonding curve rate of the current funding cycle.{' '}
-            <span style={{ fontWeight: 500, color: colors.text.warn }}>
-              Tokens are burned when they are redeemed.
-            </span>
-          </p>
-          {redeemDisabled && (
-            <div style={{ color: colors.text.secondary, fontWeight: 500 }}>
-              You can redeem tokens once this project has overflow.
-            </div>
-          )}
-          {!redeemDisabled && (
-            <div>
-              <FormattedNumberInput
-                min={0}
-                step={0.001}
-                placeholder="0"
-                value={redeemAmount}
-                disabled={redeemDisabled}
-                accessory={
-                  <InputAccessoryButton
-                    content="MAX"
-                    onClick={() => setRedeemAmount(fromWad(totalBalance))}
-                  />
-                }
-                onChange={val => setRedeemAmount(val)}
-              />
-              <div style={{ fontWeight: 500, marginTop: 20 }}>
-                You will receive minimum {formatWad(rewardAmount) || '--'} ETH
-              </div>
-            </div>
-          )}
-        </Space>
-      </Modal>
-      {/* <ConfirmStakeTokensModal
-        visible={stakeModalVisible}
-        onCancel={() => setStakeModalVisible(false)}
-        ticketsUpdateOn={ticketsUpdateOn}
-      /> */}
+      />
       <ConfirmUnstakeTokensModal
         visible={unstakeModalVisible}
         onCancel={() => setUnstakeModalVisible(false)}
